fix error showing (#15)
